Move function calling agent config into its own folder
